server/requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.3
django-allauth==0.61.1
django-cors-headers==4.3.1
django_crontab==0.7.1
django-dbbackup==4.1.0
django-env-overrides==0.1.0
django-environ==0.11.2
django-filter==23.3
django-import-export==3.3.7
django-jazzmin==2.6.0
django-jsonfield==1.4.1
django-mass-edit==3.5.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.3
MarkupPy==1.14
MarkupSafe==2.1.3
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
openapi-codec==1.3.2
opencv-python==4.10.0.84
openpyxl==3.1.2
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyotp==2.9.0
Python-IO==0.3
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.4.0
scipy==1.12.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tzlocal==5.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
